separated graph classes
