+++ include/zcalc/graph/vertex.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <cstdint>
+
+namespace zcalc {
+namespace graph {
+
+typedef std::uint32_t Vertex;
+
+} /* namespace graph */
+} /* namespace zcalc */
